requirements.txt
langchain==0.1.20
langchain-openai==0.1.0
langchain-chroma==0.1.0
langchain-community==0.0.38
chromadb==0.4.22
pyyaml==6.0.1
requests==2.31.0
python-dotenv==1.0.0
gradio==4.44.1
pytest==7.4.3